chore: upgrade deps
